feat: update unit test
